chore: enable auto code splitting + fix incorrect path
