+++ .storybook/preview.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { ThemeProvider } from 'styled-components';
 import { darkTheme } from '../packages/lib-components/src/styled';
 import { GlobalStyle } from '../packages/lib-components/src/styled/GlobalStyle';
-import { StepperProvider } from '../packages/lib-components/src/components/navigation/Steppers/context';
+import { StepperProvider } from '../packages/lib-components/src/components/navigation/Steppers/provider';
 import { SnackbarProvider } from '../packages/lib-components/src/helpers/getSnackbarProvider';
 import 'simplebar-react/dist/simplebar.min.css';
 import { Preview } from '@storybook/react';
