+++ packages/lib-components/package.json
@@ -31,7 +31,7 @@
     "@rjsf/validator-ajv8": "5.20.0",
     "@sentry/react": "8.53.0",
     "@tanstack/react-table": "8.20.5",
-    "@tanstack/react-router": "1.79.0",
+    "@tanstack/react-router": "1.102.1",
     "@types/luxon": "3.4.2",
     "@visx/axis": "3.10.1",
     "@visx/brush": "3.10.4",
