+++ packages/web-main/package.json
@@ -11,9 +11,9 @@
     "@sentry/tracing": "7.120.3",
     "@tanstack/react-query": "5.56.2",
     "@tanstack/react-query-devtools": "5.58.0",
-    "@tanstack/react-router": "1.79.0",
-    "@tanstack/router-devtools": "1.79.0",
-    "@tanstack/router-vite-plugin": "1.79.0",
+    "@tanstack/react-router": "1.102.1",
+    "@tanstack/router-devtools": "1.102.1",
+    "@tanstack/router-vite-plugin": "1.102.1",
     "luxon": "3.5.0",
     "react-dnd": "16.0.1",
     "react-dnd-html5-backend": "16.0.1",
