+++ packages/web-main/vite.config.mts
@@ -20,7 +20,7 @@ export default defineConfig({
     warmup: {
       clientFiles: [
         // commonly used files to warm up the esbuild cache
-        './src/Router.tsx',
+        './src/router.tsx',
         './src/index.tsx',
       ],
     },
@@ -28,7 +28,7 @@ export default defineConfig({
   css: {
     transformer: 'lightningcss',
   },
-  plugins: [react(), TanStackRouterVite(), checker({ typescript: true }), tsconfigPaths()],
+  plugins: [react(), TanStackRouterVite({ autoCodeSplitting: true }), checker({ typescript: true }), tsconfigPaths()],
 
   envDir: '../../',
 });
